Compilation failure
This change results in an inlined implementation of the write()
function that does not check for termination, seemingly confused
by the ensuing infinite loop.

The result is that the NUL terminator of the string is also written
out into the TX FIFO and attempting to read a subsequent byte will
result in a CPU failure (Illegal instruction reported repeatedly).

The disassembly shows that what would ordinarily have been a
'bnez <>' instruction has been replaced with a 'j <>' instruction
and the inlined loop from write() continues executing beyond the
point at which it should have terminated.