Create SD card test
SD card test checks for the presence of a microSD card in the
slot, reads and reports the card properties and then proceeds
to check the contents of a test file `LORUM.IPS` in the root
directory of a FAT32-formatted card against its internal
reference.

See the source for directions on how to set up a microSD card
with a suitable file; set 'emitText' to true on the first run
and it will emit the sample text with instructions.