Iniitial attempts to bring up debugger
Modify tcp_server to check for available buffer space before
receiving data.
Introduce reset of JTAG TAP in TestLogicReset state; this is
already done in the OT project but was not present in this
implementation. The TestLogicReset state must reset the
internal logic and drive out a reset signal. Without this reset
the IDCODE typically reads incorrectly.
Ensure device_err signal is driven.