Modify the debug module to support CHERIoT and non-CHERIoT debugging.
Add a simple boot sign-on program for legacy mode, because we cannot
use the CHERI code and the processor must be running for the debugger
to halt the HART and download code.

Introduce a second debug_rom implementation that provides operation
for vanilla RISC-V mode.

TODO: Tidy up the Makefile!