Introduce arbitration and switch to dual port implementation
Separated I and D ports, each with its own internal read buffer,
for increased performance. It's worth noting that the instruction
port does not require capability tag supports or write support
so the cost is appreciably less than doubling resources.

A combinatorial loop would exist between the Instruction Fetch
and LSU ports of the Ibex with the HyperRAM presented to both
ports. Introduce a single-entry FIFO on each TL-UL connection
to eliminate that loop, but use PassThrough behaviour to avoid
adding latency.

Vivado 2022.2 very much appears to have an issue with packed
arrays of structs and/or module arrays, so reluctantly a generate
block is now used for the port submodules.