Update HyperRAM test software to reflect size changes
Introduce a test for whether storing a capability to an untagged
area of memory raises an exception as intended. Stores to areas
of the HyperRAM that have associated tag bit storage should
complete without an exception.

In each case, check whether an exception does/does not occur as
expected, and check the contents of the memory after the attempted
store.

An exception handler resumes execution after the faulting
instruction but - importantly - is activated only for the very
short time window during which we expect the possible exception to
occur.